feat(modules): `db`, `real`, `song` created.
Mostly created the new modules for ease-of-use.
`real` is for things not related to the core algorithm.
`song` is for exactly that: the core algo/functioning.
Lastly, `db` is for interactions with the database.
The modules mostly serve to make it easier to focus on
the relevant parts of the code. Might add stuff to `real`.