feat(playback): user commands when playing songs.
You can now, through `-t`, play the first three songs from the database.
When playing these songs, the user can also pause, skip,
or cancel (stop). The updated vector containing the songs (so with the
added touches and skips) will be displayed — **however**, this
information is not yet added to the database (as it is just 'testing').